Copy my existing code to the new repository and add a license (GPL-3.0) + make the README more useful.
My old repo was a standalone bot and had lots of unrelated code from that as well as multiple tokens and keys that should really never have been committed (not that the codebase was publically accessible, but regardless...). I'm copying this all over to a new repo that can safely be made public, and adding a license so that people can use it.